Commented potential fix for 16bpp-Blue-byte-bug
HzMod seems to be sending bad data for the blue byte in Super Mario 3D Land. For that, we can use an alternate formula (which I have documented and tested), but this formula produces incorrect results for Pokemon Yellow. I don't know exactly why this happens. More investigation is needed. It's also very possible that HzMod ***is*** broken in this regard, at least sometimes, so maybe I'll work on that end of the exchange...